Refactor authentication on GNU/Linux.
Previously authentication was done by running multibootusb through gksudo or
saving the password manually and piping it in to sudo. This approach had a few
problems:

* Running as root disallows user modification.
* Running GUI programs as root is a bad idea.
* Qt4 (and so this program) won't run as root any more. [1]
* Caching passwords bypasses secure GUI authentication.

The solution I've come up with is to bite the bullet and run multibootusb as a
user and use gksudo/gksu/kdesudo/kdesu to authenticate properly.

The most major change done here is writing some psuedo-shell scripts to be
passed to the new admin.adminCmd program, since prompting multiple times is
annoying.

This commit breaks the ISO imager, it will be fixed in the next commit.

[1]: https://bugzilla.redhat.com/show_bug.cgi?id=1212180#c7