Rewrite dd code to run through sudo on GNU/Linux.
The last commit breaks the ability to send signals or stdin to applications
running as root, and unfortunately reading dd's status requires that.

One way to fix this would be to run dd as root with a program like 'pv' to
output progress. This works, but has the downside that you can't cancel the
process if multibootusb is closed, it just keeps running as root.

The solution here is to use a named pipe between multibootusb and dd, and send
the ISO through that. The advantage is that the bytes written can be directly
calculated. If dd closes the pipe, it can be handled, and if we close the pipe
dd will handle it. No processes will lay around idly eating disks.