Throw exception when partition is not mounted.
Instead of returning strings for partition size value,
throw an exception when the partition is not (and presumably cannot be) mounted
in `details_udisks2()`.

Also, return 0s instead of empty strings in a similar situation in `details_udev()`

This changes prevent the conversion error reported in issue #397; however,
additional changes will be needed to catch the `PartitionNotMounted` exception
now potentially thrown by a call to `details()` on Linux.